requirements.txt
telegram
configparser
redis
mysql-connector